Implemented getNsimilarUsers for RecommendationDataAccessInterface

+++ src/data_access/similarityMapUpdaterFacade/mapGenerator/APICallerInterface.java
@@ -1,5 +1,5 @@
 package data_access.similarityMapUpdaterFacade.mapGenerator;
 
 public interface APICallerInterface {
-    public Float getSimilarityScore(String text1, String text2);
+    public Double getSimilarityScore(String text1, String text2);
 }
